+++ android/app/src/main/java/chi/epsilon/epsilon/boozebot/activities/LoginActivity.java
@@ -1,4 +1,4 @@
-package chi.epsilon.epsilon.boozebot;
+package chi.epsilon.epsilon.boozebot.activities;
 
 import android.support.v4.app.Fragment;
 import android.support.v4.app.FragmentManager;
@@ -6,6 +6,9 @@ import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
 
 import com.firebase.client.Firebase;
+
+import chi.epsilon.epsilon.boozebot.fragments.MainLoginFragment;
+import chi.epsilon.epsilon.boozebot.R;
 
 public class LoginActivity extends AppCompatActivity {
