+++ android/app/src/main/java/chi/epsilon/epsilon/boozebot/fragments/MainLoginFragment.java
@@ -1,4 +1,4 @@
-package chi.epsilon.epsilon.boozebot;
+package chi.epsilon.epsilon.boozebot.fragments;
 
 import android.os.Bundle;
 import android.support.v4.app.Fragment;
@@ -14,7 +14,8 @@ import com.firebase.client.AuthData;
 import com.firebase.client.Firebase;
 import com.firebase.client.FirebaseError;
 
-import java.util.Map;
+import chi.epsilon.epsilon.boozebot.R;
+import chi.epsilon.epsilon.boozebot.fragments.CreateAccountFragment;
 
 public class MainLoginFragment extends Fragment {
     private Button mLoginButton;
